Add ruby and gem configuration to replicate GH pages
This adds the `github-pages` gem, which attempts to replicate the GH
pages build process locally. It seems to work, and replicate production
quite well.

The one difference here is that GH pages uses a quite old version of
ruby (2.7), which was EOL in March 2023. So, let's use the most recent
Ruby and just deal with any incompatibilities.