Use relative_url and baseurl to standardize GH pages URLs
The assets were broken since locally they were generated in `/assets`
but on GitHub, it's actually under
`/FFS-IncomeVerificationToolkit/assets` due to the GH page structure.

There are a million people asking on StackOverflow about this over the
years, and the best advice seems to be to just set the "baseurl" to your
repo name and using `relative_url` to prepend the baseurl.